requirements.txt
flask>=2.3.0
openai-whisper>=20231117
deep-translator>=1.11.4
gtts>=2.5.1
